Move armconversion package up one level
