+++ hack/generator/pkg/armconversion/arm_conversion_function.go
