+++ hack/generator/pkg/armconversion/create_empty_arm_value_function.go
