+++ hack/generator/pkg/armconversion/shared.go
